requirements.txt
annotated-types==0.6.0
av==11.0.0
ffmpeg-python==0.2.0
future==1.0.0
passlib==1.7.4
pillow==10.2.0
psycopg==3.1.18
pydantic==2.6.3
pydantic_core==2.16.3
pyfakefs==5.5.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.10.0